Revert "Feat/delete document.usecase"
