+++ backend/prisma/schema.prisma
@@ -29,9 +29,6 @@ model User {
   teacherProfile TeacherProfile? @relation("UserTeacherProfile")
   studentProfile StudentProfile? @relation("UserStudentProfile")
 
-  // Documentos subidos
-  documents Document[]
-
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
 }
@@ -68,7 +65,6 @@ model UserRole {
   roleId String
   user   User   @relation(fields: [userId], references: [id], onDelete: Cascade)
   role   Role   @relation(fields: [roleId], references: [id], onDelete: Cascade)
-
 
   @@id([userId, roleId])
   @@index([roleId])
@@ -145,156 +141,8 @@ model Document {
   s3Key        String
   size         Int
   contentType  String
-  fileHash         String                    @unique // Hash SHA-256 del archivo
-  extractedText    String?                   @db.Text // Texto extraído
-  status           DocumentStatus            @default(UPLOADED)
-  
-  // Metadata del documento
-  uploadedBy       String
-  uploader         User                      @relation(fields: [uploadedBy], references: [id])
-  
-  // Relations para RAG
-  chunks           DocumentChunk[]
-  categories       DocumentCategoryMapping[]
-  processingJobs   ProcessingJob[]
-  
-  // Información adicional extraída
-  pageCount        Int?                      // Número de páginas (PDFs)
-  documentTitle    String?                   // Título extraído del documento
-  documentAuthor   String?                   // Autor extraído
-  language         String?                   // Idioma detectado
-
   uploadedAt   DateTime @default(now())
-  updatedAt        DateTime                  @updatedAt
-  
-  @@index([status])
-  @@index([uploadedBy])
-  @@index([fileHash])
-  @@index([contentType])
 }
-
-model DocumentChunk {
-  id              String    @id @default(uuid())
-  documentId      String
-  document        Document  @relation(fields: [documentId], references: [id], onDelete: Cascade)
-  
-  content         String    @db.Text         // Contenido del chunk
-  chunkIndex      Int                        // Índice del chunk en el documento
-  startPosition   Int                        // Posición de inicio en el texto original
-  endPosition     Int                        // Posición de fin en el texto original
-  pageNumber      Int?                       // Página de origen (si aplica)
-  type            String    @default("text") // Tipo de chunk (text, heading, list, etc.)
-  
-  // Metadata estadística
-  wordCount       Int       @default(0)      // Número de palabras
-  charCount       Int       @default(0)      // Número de caracteres
-  
-  // pgvector para embeddings
-  embedding       Unsupported("public.vector(1536)")? // OpenAI embeddings (1536 dimensiones)
-  
-  // Metadata adicional del chunk
-  metadata        Json?                      // Información contextual adicional
-  
-  createdAt       DateTime  @default(now())
-  updatedAt       DateTime  @updatedAt
-  
-  @@unique([documentId, chunkIndex])
-  @@index([documentId])
-  @@index([type])
-  @@index([wordCount])
-  // Índice HNSW para búsqueda vectorial eficiente
-  // CREATE INDEX CONCURRENTLY document_chunks_embedding_hnsw_idx ON "DocumentChunk" USING hnsw (embedding vector_cosine_ops);
-  @@map("document_chunks")
-}
-
-model DocumentCategory {
-  id          String                      @id @default(uuid())
-  name        String                      @unique
-  description String?
-  color       String?                     // Color para UI
-  icon        String?                     // Icono para UI
-  
-  documents   DocumentCategoryMapping[]
-  
-  createdAt   DateTime                    @default(now())
-  updatedAt   DateTime                    @updatedAt
-}
-
-model DocumentCategoryMapping {
-  documentId   String
-  categoryId   String
-  document     Document         @relation(fields: [documentId], references: [id], onDelete: Cascade)
-  category     DocumentCategory @relation(fields: [categoryId], references: [id], onDelete: Cascade)
-  
-  createdAt    DateTime         @default(now())
-  
-  @@id([documentId, categoryId])
-}
-
-model ProcessingJob {
-  id           String           @id @default(uuid())
-  documentId   String
-  document     Document         @relation(fields: [documentId], references: [id], onDelete: Cascade)
-  
-  jobType      ProcessingType
-  status       ProcessingStatus @default(PENDING)
-  errorMessage String?          @db.Text
-  progress     Int              @default(0)     // Progreso 0-100
-  
-  // Detalles específicos del job
-  jobDetails   Json?            // Configuración específica del job
-  result       Json?            // Resultado del procesamiento
-  
-  startedAt    DateTime?
-  completedAt  DateTime?
-  createdAt    DateTime         @default(now())
-  
-  @@index([documentId, jobType])
-  @@index([status])
-  @@index([createdAt])
-}
-
-/**
- * ===========================
- * Enums para Document Processing
- * ===========================
- */
-
-enum DocumentStatus {
-  UPLOADED      // Archivo subido, esperando procesamiento
-  PROCESSING    // En proceso de extracción/chunking/embedding
-  PROCESSED     // Completamente procesado y listo para búsqueda
-  ERROR         // Error en el procesamiento
-  DELETED       // Marcado como eliminado
-}
-
-enum ProcessingType {
-  TEXT_EXTRACTION       // Extraer texto del archivo
-  CHUNKING             // Dividir texto en chunks
-  EMBEDDING_GENERATION // Generar embeddings vectoriales
-  FULL_PROCESSING      // Procesamiento completo (todos los pasos)
-  REPROCESSING         // Reprocesar documento existente
-}
-
-enum ProcessingStatus {
-  PENDING      // Esperando en cola
-  RUNNING      // Ejecutándose actualmente
-  COMPLETED    // Completado exitosamente
-  FAILED       // Falló con error
-  CANCELLED    // Cancelado manualmente
-  RETRYING     // Reintentando después de fallo
-}
-
-enum ChunkType {
-  TEXT         // Chunk de texto normal
-  TITLE        // Título o encabezado
-  TABLE        // Contenido de tabla
-  LIST         // Lista o enumeración
-  CODE         // Código o texto técnico
-  FORMULA      // Fórmula matemática
-  METADATA     // Metadatos del documento
-}
-
 /**
  * ===========================
  * Exams generation
