+++ backend/src/modules/repositoryDocuments/application/commands/delete-document.usecase.ts
