+++ backend/src/modules/repositoryDocuments/domain/ports/document-storage.port.ts
@@ -10,11 +10,4 @@ export interface DocumentStoragePort {
   listDocuments(): Promise<DocumentListItem[]>;
   documentExists(fileName: string): Promise<boolean>;
   softDeleteDocument(fileName: string): Promise<void>;
-
-  /**
-   * Descarga el contenido de un archivo como Buffer
-   * @param fileName Nombre del archivo o clave S3
-   * @returns Buffer con el contenido del archivo
-   */
-  downloadFileBuffer(fileName: string): Promise<Buffer>;
 }
