+++ backend/src/modules/repositoryDocuments/tokens.ts
@@ -1,0 +1,2 @@
+export const DOCUMENT_REPO = Symbol('DocumentRepositoryPort');
+export const FILE_STORAGE_REPO = Symbol('FileStorageRepositoryPort');
