+++ client/src/interfaces/documentInterface.ts
@@ -2,9 +2,6 @@ export interface Document {
   fileName: string;
   originalName: string;
   mimeType: string;
-  fileName: string;
-  originalName: string;
-  mimeType: string;
   size: number;
   downloadUrl: string;
   uploadedAt: string;
