+++ infra/docker/compose.dev.yml
@@ -1,6 +1,6 @@
 services:
   db:
-    image: pgvector/pgvector:pg16
+    image: postgres:16-alpine
     container_name: learning-agent-db
     restart: unless-stopped
     ports:
